+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Snippet [autocompletion](./README.md#autocompletion) for the Lua mode
 - Additional [lint](./README.md#lint) rules for the `sanitized-css` pages on a MediaWiki site
+- Smart indentation for JavaScript, CSS, Lua and JSON modes
 
 ## 2.27.2